[FEATURE] New database

+++ src/models/nginx_log.rs
@@ -11,7 +11,8 @@ fn deserialize_empty_string<'de, D>(
   deserializer: D,
 ) -> Result<Option<String>, D::Error>
 where
-  D: Deserializer<'de>, {
+  D: Deserializer<'de>,
+{
   let buf = String::deserialize(deserializer)?;
   if buf.is_empty() {
     Ok(None)
@@ -22,7 +23,8 @@ where
 
 fn deserialize_string_to_i64<'de, D>(deserializer: D) -> Result<i64, D::Error>
 where
-  D: Deserializer<'de>, {
+  D: Deserializer<'de>,
+{
   let buf = String::deserialize(deserializer)?;
   let res = buf.parse::<i64>().unwrap_or_default();
   Ok(res)
@@ -30,7 +32,8 @@ where
 
 fn deserialize_string_to_i32<'de, D>(deserializer: D) -> Result<i32, D::Error>
 where
-  D: Deserializer<'de>, {
+  D: Deserializer<'de>,
+{
   let buf = String::deserialize(deserializer)?;
   let res = buf.parse::<i32>().unwrap_or_default();
   Ok(res)
@@ -38,7 +41,8 @@ where
 
 fn deserialize_string_to_f64<'de, D>(deserializer: D) -> Result<f64, D::Error>
 where
-  D: Deserializer<'de>, {
+  D: Deserializer<'de>,
+{
   let buf = String::deserialize(deserializer)?;
   let res = buf.parse::<f64>().unwrap_or_default();
   Ok(res)
@@ -55,8 +59,8 @@ pub struct NginxLogPartial {
   pub(crate) request_method: String,
   #[serde(deserialize_with = "deserialize_string_to_i64")]
   pub(crate) content_length: i64,
-  #[serde(deserialize_with = "deserialize_string_to_i32")]
-  pub(crate) status: i32,
+  #[serde(deserialize_with = "deserialize_string_to_i64")]
+  pub(crate) status: i64,
   #[serde(deserialize_with = "deserialize_string_to_f64")]
   pub(crate) request_time: f64,
   #[serde(deserialize_with = "deserialize_string_to_i64")]
@@ -92,7 +96,7 @@ pub struct NginxLogItem {
   pub(crate) server_protocol: String,
   pub(crate) request_method: String,
   pub(crate) content_length: i64,
-  pub(crate) status: i32,
+  pub(crate) status: i64,
   pub(crate) request_time: f64,
   pub(crate) body_bytes_sent: i64,
   pub(crate) proxy_host: Option<String>,
